Slight Refactoring
