+++ Config.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel;
-using System.Collections.Generic;
 using Terraria.ModLoader;
 using Terraria.ModLoader.Config;
