+++ description.txt
@@ -1,84 +1,29 @@
 Level+ is meant to bring a simple leveling system that makes your character progress as the game progresses.
 
+Wanna join a server and interact with the dev? Click here: https://discord.gg/DEwwnbW
+
 Check out what I'm working on and to come here: https://trello.com/b/jAdgFhfR/level
+
 Ideas or Bugs? Send them here: https://forms.gle/BvTxWhWfqEcyCW2o9
 
 Play expert or higher for an added challenge of enemies scaling with you.
 
 You reset stats by making a Stat Token, obtainable after killing Wall of Flesh, made from a stat emblem, crystal shards, and a life crystal.
 
-Per Level Bonuses
-	+2 life per level
-	+1 mana per level
-
-Combat Ring (The colored one):
-
-	Constitution:
-		+5 life per point
-		+1 life regen per 20 points
-		+1 defense per 3 points
-
-	Intelligence:
-		+1% magic damage per point
-		+1% magic crit chance per 15 points
-
-	Strength:
-		+1% melee damage per point
-		+1% melee crit chance per 15 points
-
-	Dexterity:
-		+1% ranged damage per point
-		+1% ranged crit chance per 15 points
-
-	Charisma:
-		+1% minion damage per point
-		+1% minion crit chance per 15 points
-	
-Utility Ring (The uncolored one):
-
-	Animalia:
-		+2% fishing skill per point
-		+1 minion per 20 points
-		+2% minion knockback per point
-
-	Excavation:
-		+1% pick speed per point
-		+2% place speed per point
-		+1 block range per 10 points
-
-	Mobility:
-		+1% max run speed per point
-		+2% acceleration per point
-		+2% flight time per point
-
-	Luck:
-		+1% xp per point
-		+1% chance not to consume ammo per point
-
-	Mysticism:
-		+2 max mana per point
-		+1 mana regen per 15 points
-		-0.5% mana cost per point (100% cap)
-
-Amount of XP gained from an enemy:
-Normal enemy: 1/3 of its life
-Boss enemy: 1/4 of its life
-
-Enemy Scaling:
-+2.5% life and damage per average level (average level is the average level of online players)
-
+All other things are configurable.
 
 Credits:
 
 Programming:
-	PoctorDepper
+- PoctorDepper
+- LanToaster with a few hotfixes and config
 
 Art:
-	PoctorDepper
+ - PoctorDepper
 
 Testing:
-	Circul1
-	Mageblood
-	AkantuS
+ - Circul1
+ - Mageblood
+ - AkantuS
 
 Special thanks to direwolf420 for helping out a lot, and Mageblood for ideas and fixes. (even though they didn't have to)
